Move cookie parse to main utils folder
