+++ internals/chat/client.py
@@ -42,13 +42,11 @@ from internals.chat.parser import (
     ChatDetails,
     YoutubeChatParser,
     complex_walk,
-    parse_expiry_as_date,
     parse_initial_data,
-    parse_netscape_cookie_to_morsel,
     parse_youtube_video_data,
 )
 from internals.chat.utils import camel_case_split, remove_prefixes, remove_suffixes, try_get_first_key
-from internals.utils import find_cookies_file
+from internals.utils import find_cookies_file, parse_expiry_as_date, parse_netscape_cookie_to_morsel
 
 if TYPE_CHECKING:
     from internals.chat.writer import JSONWriter
